user_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
loguru



